+++ readium/shared/r2-shared/src/main/java/org/readium/r2/shared/extensions/Exception.kt
@@ -9,6 +9,8 @@
 
 package org.readium.r2.shared.extensions
 
+import timber.log.Timber
+
 /**
  * Returns the result of the given [closure], or null if an [Exception] was raised.
  */
@@ -21,3 +23,14 @@ inline fun <T> tryOrNull(closure: () -> T): T? =
 inline fun <T> tryOr(default: T, closure: () -> T): T =
     try { closure() }
     catch (e: Exception) { default }
+
+/**
+ * Returns the result of the given [closure], or null if an [Exception] was raised.
+ * The [Exception] will be logged.
+ */
+inline fun <T> tryOrLog(closure: () -> T): T? =
+    try { closure() }
+    catch (e: Exception) {
+        Timber.e(e)
+        null
+    }