+++ readium/shared/r2-shared/src/main/java/org/readium/r2/shared/extensions/JSON.kt
@@ -132,11 +132,17 @@ fun JSONObject.optPositiveDouble(name: String, fallback: Double = -1.0, remove: 
 }
 
 /**
- * Returns the value mapped by [name] if it exists, coercing it if necessary, or [null] if no such
+ * Returns the value mapped by [name] if it exists, coercing it if necessary, or `null` if no such
  * mapping exists.
  * If [remove] is true, then the mapping will be removed from the [JSONObject].
  */
 fun JSONObject.optNullableString(name: String, remove: Boolean = false): String? {
+    // optString() returns "null" if the key exists but contains the `null` value.
+    // https://stackoverflow.com/questions/18226288/json-jsonobject-optstring-returns-string-null
+    if (isNull(name)) {
+        return null
+    }
+
     val string = optString(name)
     val value = if (string != "") string else null
     if (remove) {
@@ -146,7 +152,7 @@ fun JSONObject.optNullableString(name: String, remove: Boolean = false): String?
 }
 
 /**
- * Returns the value mapped by [name] if it exists, coercing it if necessary, or [null] if no such
+ * Returns the value mapped by [name] if it exists, coercing it if necessary, or `null` if no such
  * mapping exists.
  * If [remove] is true, then the mapping will be removed from the [JSONObject].
  */
@@ -162,7 +168,7 @@ fun JSONObject.optNullableBoolean(name: String, remove: Boolean = false): Boolea
 }
 
 /**
- * Returns the value mapped by [name] if it exists, coercing it if necessary, or [null] if no such
+ * Returns the value mapped by [name] if it exists, coercing it if necessary, or `null` if no such
  * mapping exists.
  * If [remove] is true, then the mapping will be removed from the [JSONObject].
  */
@@ -178,7 +184,7 @@ fun JSONObject.optNullableInt(name: String, remove: Boolean = false): Int? {
 }
 
 /**
- * Returns the value mapped by [name] if it exists, coercing it if necessary, or [null] if no such
+ * Returns the value mapped by [name] if it exists, coercing it if necessary, or `null` if no such
  * mapping exists.
  * If [remove] is true, then the mapping will be removed from the [JSONObject].
  */
@@ -194,7 +200,7 @@ fun JSONObject.optNullableLong(name: String, remove: Boolean = false): Long? {
 }
 
 /**
- * Returns the value mapped by [name] if it exists, coercing it if necessary, or [null] if no such
+ * Returns the value mapped by [name] if it exists, coercing it if necessary, or `null` if no such
  * mapping exists.
  * If [remove] is true, then the mapping will be removed from the [JSONObject].
  */
@@ -229,7 +235,7 @@ fun JSONObject.optStringsFromArrayOrSingle(name: String, remove: Boolean = false
 /**
  * Returns a list containing the results of applying the given transform function to each element
  * in the original [JSONObject].
- * If the tranform returns [null], it is not included in the output list.
+ * If the tranform returns `null`, it is not included in the output list.
  */
 fun <T> JSONObject.mapNotNull(transform: (Pair<String, Any>) -> T?): List<T> {
     val result = mutableListOf<T>()
@@ -245,7 +251,7 @@ fun <T> JSONObject.mapNotNull(transform: (Pair<String, Any>) -> T?): List<T> {
 /**
  * Returns a list containing the results of applying the given transform function to each element
  * in the original [JSONArray].
- * If the tranform returns [null], it is not included in the output list.
+ * If the tranform returns `null`, it is not included in the output list.
  */
 fun <T> JSONArray.mapNotNull(transform: (Any) -> T?): List<T> {
     val result = mutableListOf<T>()